+++ src/compnents/organisms/RegisterCard.tsx
@@ -1,38 +1,20 @@
+import { VFC } from "react";
 import { FormControl, FormLabel } from "@chakra-ui/form-control";
 import { Input } from "@chakra-ui/input";
 import { Box, Stack, Text } from "@chakra-ui/layout";
 import { Textarea } from "@chakra-ui/textarea";
-import { useRef, useState, VFC } from "react";
+
+import { useBookRegister } from "../../hooks/useBookRegister";
 import { MainButton } from "../atoms/MainButton";
-import { collection, addDoc } from "firebase/firestore";
-import { firebaseStore } from "../../config/firebase";
-import { useHistory } from "react-router";
-// import axios from "axios"; axiosに実装変換予定
 
 export const RegisterCard: VFC = () => {
-  const [isLoading, setLoading] = useState(false);
-  const titleRef = useRef<HTMLInputElement>(null);
-  const articleRef = useRef<HTMLInputElement>(null);
-  const descriptionRef = useRef<HTMLTextAreaElement>(null);
-  const history = useHistory();
-  const onClickRegister = async () => {
-    try {
-      setLoading(true);
-      const docRef = await addDoc(collection(firebaseStore, "books"), {
-        title: titleRef.current?.value,
-        article: articleRef.current?.value,
-        description: descriptionRef.current?.value,
-        image: "https://source.unsplash.com/random",
-      });
-      console.log("Document written with ID: ", docRef.id);
-      alert("登録しました");
-      setLoading(false);
-      history.push("/");
-    } catch (error) {
-      alert(error);
-      setLoading(false);
-    }
+  const { isLoading, titleRef, articleRef, descriptionRef, registBookInfo } =
+    useBookRegister();
+
+  const onClickRegister = () => {
+    registBookInfo();
   };
+
   return (
     <Box
       minW="600px"
