+++ src/types/BookType.ts
@@ -1,0 +1,7 @@
+export type BookType = {
+  id: string;
+  title: string;
+  article: string;
+  description: string;
+  image: string;
+};
